+++ python/transitio/edit/__init__.py
@@ -2,6 +2,7 @@
 
 from transitio.edit._build import build_feed
 from transitio.edit._editor import FeedBuilder, FeedEditor
+from transitio.edit._osm import OsmEditor
 from transitio.edit._snap import snap_to_network
 
-__all__ = ["FeedBuilder", "FeedEditor", "build_feed", "snap_to_network"]
+__all__ = ["FeedBuilder", "FeedEditor", "OsmEditor", "build_feed", "snap_to_network"]